+++ contracts/interfaces/ILendingMarketCapybaraV1.sol
@@ -28,21 +28,27 @@ interface ILendingMarketCapybaraV1 {
     ) external returns (uint256);
 
     /**
-     * @dev Takes an installment loan with multiple sub-loans for a provided account.
+     * @dev Takes an installment loan with multiple sub-loans for a provided account with additional parameters.
+     *
+     * See notes about the penalty interest rate in the CapybaraFinance V1 repository
+     * https://github.com/cloudwalk/brlc-capybara-finance
+     *
      * @param borrower The account for whom the loan is taken.
      * @param programId The identifier of the program to take the loan from.
-     * @param borrowAmounts The desired amounts of tokens to borrow for each installment.
+     * @param borrowedAmounts The desired amounts of tokens to borrow for each installment.
      * @param addonAmounts The off-chain calculated addon amounts for each installment.
      * @param durationsInPeriods The desired duration of each installment in periods.
+     * @param penaltyInterestRates The penalty interest rates for each installment.
      * @return firstInstallmentId The unique identifier of the first sub-loan of the installment loan.
      * @return installmentCount The total number of installments.
      */
-    function takeInstallmentLoanFor(
+    function takeInstallmentLoan(
         address borrower,
         uint32 programId,
-        uint256[] calldata borrowAmounts,
+        uint256[] calldata borrowedAmounts,
         uint256[] calldata addonAmounts,
-        uint256[] calldata durationsInPeriods
+        uint256[] calldata durationsInPeriods,
+        uint256[] calldata penaltyInterestRates
     ) external returns (uint256 firstInstallmentId, uint256 installmentCount);
 
     /**